Rename `SnapTarget` to `KeyframeSnapTarget`
